fix(test): harden preset node boundaries
