+++ packages/test/tests/android-nodes.test.ts
@@ -18,8 +18,15 @@ describe('createAndroidNodes', () => {
   it('runs ADB shell commands and maps timeoutMs to the Agent option', async () => {
     const runAdbShell = vi.fn(async () => 'package:com.example.app');
     const registry = new NodeRegistry(
-      createAndroidNodes({ getAgent: () => ({ runAdbShell }) }),
+      createAndroidNodes({
+        getAgent: () => ({
+          launch: vi.fn(async () => undefined),
+          terminate: vi.fn(async () => undefined),
+          runAdbShell,
+        }),
+      }),
     );
+    expect(registry.names()).toEqual(['launch', 'terminate', 'runAdbShell']);
     const result = await runCollectedCase(
       collected([
         {
@@ -43,7 +50,13 @@ describe('createAndroidNodes', () => {
   it('supports string shorthand and rejects adb-prefixed commands', async () => {
     const runAdbShell = vi.fn(async () => 'ok');
     const registry = new NodeRegistry(
-      createAndroidNodes({ getAgent: () => ({ runAdbShell }) }),
+      createAndroidNodes({
+        getAgent: () => ({
+          launch: vi.fn(async () => undefined),
+          terminate: vi.fn(async () => undefined),
+          runAdbShell,
+        }),
+      }),
     );
     const shorthand = await runCollectedCase(
       collected([
@@ -73,12 +86,45 @@ describe('createAndroidNodes', () => {
     );
   });
 
+  it('owns and delegates Android lifecycle nodes', async () => {
+    const launch = vi.fn(async () => undefined);
+    const terminate = vi.fn(async () => undefined);
+    const registry = new NodeRegistry(
+      createAndroidNodes({
+        getAgent: () => ({
+          launch,
+          terminate,
+          runAdbShell: vi.fn(async () => ''),
+        }),
+      }),
+    );
+    const result = await runCollectedCase(
+      collected([
+        {
+          node: 'launch',
+          input: { uri: 'com.example.app' },
+          meta: { continueOnError: false },
+        },
+        {
+          node: 'terminate',
+          input: { prompt: 'com.example.app' },
+          meta: { continueOnError: false },
+        },
+      ]),
+      { resolveNode: registry.require.bind(registry), context: undefined },
+    );
+
+    expect(result.status).toBe('success');
+    expect(launch).toHaveBeenCalledWith('com.example.app');
+    expect(terminate).toHaveBeenCalledWith('com.example.app');
+  });
+
   it('validates factory options and Agent capability', async () => {
     expect(() => createAndroidNodes({} as never)).toThrow(
       'createAndroidNodes() requires getAgent()',
     );
     const registry = new NodeRegistry(
-      createAndroidNodes({ getAgent: () => ({}) }),
+      createAndroidNodes({ getAgent: () => ({}) as never }),
     );
     const result = await runCollectedCase(
       collected([
@@ -92,6 +138,20 @@ describe('createAndroidNodes', () => {
     );
     expect(result.steps[0].error?.message).toContain(
       'Android Agent with runAdbShell()',
+    );
+
+    const lifecycleResult = await runCollectedCase(
+      collected([
+        {
+          node: 'launch',
+          input: { uri: 'com.example.app' },
+          meta: { continueOnError: false },
+        },
+      ]),
+      { resolveNode: registry.require.bind(registry), context: undefined },
+    );
+    expect(lifecycleResult.steps[0].error?.message).toContain(
+      'Android Agent with launch()',
     );
   });
 });
