+++ packages/test/tests/types/test-project-config.ts
@@ -1,5 +1,8 @@
+import type { AndroidAgent } from '@midscene/android';
+import type { IOSAgent } from '@midscene/ios';
 import { defineNode, z } from '@midscene/test';
 import {
+  type AndroidRunnerAgent,
   createAndroidNodes,
   runAdbShellInputSchema,
 } from '@midscene/test/android';
@@ -9,7 +12,12 @@ import {
   defineTestProject,
   loadTestProject,
 } from '@midscene/test/config';
-import { createIOSNodes, runWdaRequestInputSchema } from '@midscene/test/ios';
+import {
+  type IOSRunnerAgent,
+  type RunWdaRequestNodeInput,
+  createIOSNodes,
+  runWdaRequestInputSchema,
+} from '@midscene/test/ios';
 import {
   clearCookiesInputSchema,
   createPlaywrightNodes,
@@ -154,12 +162,16 @@ interface PlatformContext {
   page: Page;
   baseUrl: string;
   android: {
+    launch(uri: string): Promise<void>;
+    terminate(uri: string): Promise<void>;
     runAdbShell(
       command: string,
       options?: { timeout?: number },
     ): Promise<string>;
   };
   ios: {
+    launch(uri: string): Promise<void>;
+    terminate(uri: string): Promise<void>;
     runWdaRequest(input: {
       method: 'GET' | 'POST' | 'DELETE' | 'PUT';
       endpoint: string;
@@ -181,6 +193,22 @@ createAndroidNodes<PlatformContext>({
 createIOSNodes<PlatformContext>({
   getAgent: ({ context }) => context.ios,
 });
+
+declare const androidAgent: AndroidAgent;
+declare const iosAgent: IOSAgent;
+declare const androidRunnerAgent: AndroidRunnerAgent;
+declare const iosRunnerAgent: IOSRunnerAgent;
+declare const iosAgentInput: Parameters<IOSAgent['runWdaRequest']>[0];
+declare const iosRunnerInput: RunWdaRequestNodeInput;
+
+createAndroidNodes({ getAgent: () => androidAgent });
+createIOSNodes({ getAgent: () => iosAgent });
+androidAgent satisfies AndroidRunnerAgent;
+iosAgent satisfies IOSRunnerAgent;
+androidRunnerAgent.runAdbShell satisfies AndroidAgent['runAdbShell'];
+iosRunnerAgent.runWdaRequest satisfies IOSAgent['runWdaRequest'];
+iosAgentInput satisfies RunWdaRequestNodeInput;
+iosRunnerInput satisfies Parameters<IOSAgent['runWdaRequest']>[0];
 
 void gotoUrlInputSchema;
 void setCookiesInputSchema;
